+++ src/Sulu/Bundle/TagBundle/Resources/public/js/components/tags/components/list/main.js
@@ -126,6 +126,7 @@ define(function() {
                     el: this.sandbox.dom.find('#tags-list', this.$el),
                     url: '/admin/api/tags?flat=true',
                     resultKey: 'tags',
+                    storageName: 'tags',
                     searchFields: ['name'],
                     instanceName: constants.datagridInstanceName,
                     viewOptions: {